Initial ENVI data loading code
This is basic code to load ENVI format files, including both the header
file and the binary data file, into numpy.  It has grown to the point
that it is worth putting into a repository.

The data can't be accessed yet because the ENVI data isn't wrapped into
a class, but this will be added in the future.