Raster-view and summary-view improvements
Raster-view can now zoom-to-fit in multiple modes, although only the
"fit both dimensions" mode is correct.  (The other modes need to take
the scrollbars into account, and this is not done accurately yet.)  The
result is that the summary-view now doesn't show scrollbars when "fit
both dimensions" is selected.

Raster-view can also now report the visible area of the raster data now,
in raster data-set coordinates.  Other components can register for
changes of this value, so that we can build more UI features around this
information.  The first client will be the summary-view, which needs to
show a rectangle that indicates the display region of the main-view.