Refactor UI to link displays, draw overlays
The rasterview widget now exposes functionality to receive
mouse-click/scroll/paint notifications to allow for more sophisticated
operations to be implemented.

With this functionality, the summary-view now indicates what part of the
image is shown in the main-view with a yellow rectangle.

Additionally, work was done on the data-set classes to expose more
information from the ENVI file format (bad band list, data-ignore
value).