Config file, refactored loader, spectrum plot, bad bands
Incorporated a basic config file so that we can start including
configuration details for the application.  The initial config value is
simply the data loader to use, which presently is the GDAL data loader.
This also allows the main application to be loader-agnostic, in case we
decide to add loaders for remote data, SSH data, etc. in the future.

Also plugged the spectrum-plot viewer into the application, so that we
can start seeing how that interaction will work.  This was pretty easy;
the code was already pretty modular.  But, we still need to fuss with
the matplotlib styles to get it looking acceptable.

Finally, added code to the data-set classes to filter out bad band data,
and "data ignore values" that appear in the dataset.  There are still a
few bugs to work out around this.