Migrate model into app.py, "current pixel" on detail view
The application and its internal model will be developing together, so
I migrated the model code into the app.py file for the time being.

Also added a "current pixel" highlight to the detail view.  This will
undoubtedly need to change, but it's an example of how it could look,
for now.