New dataset-chooser, new overview pane
Created a "dataset chooser" tool-button that shows a list of the
datasets, with the currently visible dataset checked.  This is a much
more compact way of allowing the user to view and change the visible
dataset.  The chooser has been included in the overview and main
image windows, but has not yet been incorporated into the zoom window.

Created a base class for dockable panes, which will make the main
application code a lot cleaner.  This abstraction is still being
developed though, and will take some time to work through the entire
app.  One of the benefits is that now these panes can be shown and
hidden via toolbar buttons and "View" menu items.  So, they can be
hidden and then re-shown.