Dataset-chooser, band-chooser, show/hide dockable panes
Note:  a number of regressions were introduced in this commit, so some
work will need to be done to restore previously working functionality.

Multiple features in this commit:

All raster-data panes now support choosing between the various datasets
loaded into the tool.  The current dataset is indicated with a check-box
in the list of options to choose from.

The zoom pane includes a basic band-chooser, allowing the user to select
whether the data should be shown in RGB or in grayscale, and what bands
should be used for each color channel.