Refactor raster-view panes; work on zoom tools
Created a RasterPane base-class for all raster-view panes so that common
code can reside in the base class.  The different usages are a bit
grungy to support, but it's way better than having duplicated code
everywhere.  It is much easier to add new features to all raster-view
panes now.

Also re-worked the zoom-level combobox so that it will display the
current zoom level, even if it's not a level in the list of options.
This required some fussing with, since the combo-box shouldn't hold on
to focus, and if it's done wrong there are infinite recursions of
events.

The raster-view panes display an error dialog-box if minimum or maximum
zoom levels are exceeded.  This behavior may need to be removed, since
it really seems to gum up the UI behavior, even to the point that the
app cannot be closed.  So, we may need to remove this.

Finally, there is still plenty of work to do on zoom behaviors; when
zooming in or out, the center of the raster-view pane should be
preserved; the image shouldn't move off in one direction or another.