Refactor raster-pane, improve raster-pane usability
"Raster-panes" are the widgets that display raster data, and also
provide toolbars for manipulating how the raster data is displayed
(zoom, bands used, etc.).

Added a pixel-selection reticle so that the pixel whose spectra is being
viewed is very clear.  With this reticle, some pixel-selection bugs were
manifested (divergence between the click location, and the pixel that
the app thought was being "selected"), so those were addressed as well.

Refactored a large chunk of the raster-pane functionality into the
base-class so that the context, main and zoom panes all benefit from the
new functionality that was added.

Reworked some of the logic controlling when the context pane is
auto-scrolled based on main-view scrolling/selection, when the main-view
is auto-scrolled based on zoom-pane scrolling/selection.  I think the
current interaction is much more user-friendly now.