Fix visible-area issue for main-view and zoom-pane
The main-view was showing the "zoom pane visible area" even when the
zoom-pane was hidden.  This was partly due to a bug in the raster-pane
code, and partly due to a weird behavior in Qt with dockable widgets.

The solution was to hook into the QWidget show/hide events, and expose
that from the raster-pane so that the application could be aware of it.
Additionally, some of the raster-pane wasn't updating the UI on some
code-paths, which resulted in the visible-area box remaining on screen.