Multiple UI/usability improvements
Added an "open image" icon to the main toolbar, so that users don't have
to use the File menu for this.

Changed the switch-image icon to match other icons in the toolbar.

Changed the dataset info pane to have scrollbars, so that it's easy to
see the full filenames of open data sets.

Changed the behavior of scrolling in the context/main views, so that
when the main view's visible-area rectangle reaches the edge of the
context pane, the context pane is re-scrolled to show the full
rectangle.

This was also done for the interaction between main/zoom views.

Still trying to figure out how to zoom in/out around the center of a
view; this bug is remaining difficult to resolve.