Fix raster views to retain image center when zooming
Addressed a major usability issue:  when zooming in or out, the
raster-view would change what part of the image was in the center of
the viewport.  This has now been fixed; the raster-view now correctly
adjusts the scrollbars to keep the image data centered in the viewport
when zooming in and out.

Hooray!!!