Change pixel-highlight reticle to be a crosshair
The initial pixel-highlight reticle was a crosshair with a hole in the
middle, but the researchers requested that it actually be a crosshair
that crosses in the middle.  This is now changed.

Additionally, added config to the application, allowing the type of
crosshair to be switched between a "small cross" and a "large cross."

There is no config UI yet, so this will need to be changed in the
code to see the different versions.