First support for color-specific stretches
Quite a bit of refactoring to support the new comboBox_affected_band
in the UI. The UI doesn't remember the settings for difference bands.
Also, StretchHistEqualize.calculate now knows how to build the needed
density histogram from the counts version used by StretchLinear.
This makes it unnecessary to get the band_data again.