Load and save ISWB project files
Implemented functionality to save ISWB configuration to a JSON file,
and to load it and restore the ISWB configuration.  Currently, the
information saved and restored is:

*   Which data files are currently loaded
*   What Regions of Interest the user has created
*   Window geometry and dock-widget/toolbar state

There are probably many other things that need to be saved and restored.
Additionally, there are a few bugs that still remain to be squashed.