Context menu for raster panes, including Regions of Interest
The Raster Pane / Raster View classes now support context-menu events.
Raster Pane has just a few simple options right now, but Regions of
Interest are included in the context menu if the click intersects the
ROI's bounding region.

In the future we will want stricter intersection tests, and all the menu
items need to be hooked up so we can do things with ROIs.