Restructure and debug stretch calculations
Factored out the old extract_band_for_display() function in
rasterview.py, since again that probably shouldn't be rasterview's
responsibility.  It also created the weird dependency between
StretchBuilder and RasterView.  Replaced it with code in dataset.py,
get_normalized_band().  This function preserves NaNs.

Factored out the application of stretch to a helper function in the
stretch.py module.  That responsibility really shouldn't be in
rasterview.py anyway, and now we can isolate and test the stretch
application functionality more easily.

Updated the StretchBuilder to use the new get_normalized_band() function
in dataset.py, so that it doesn't need to talk to the rasterview any
longer.  Now histograms don't have a peak at 0.  (Never figured out the
cause of that, but the issue is now gone.)

Factored out the application of stretch to a helper function in the
stretch.py module.  That responsibility really shouldn't be in
rasterview.py anyway, and now we can isolate and test the stretch
application functionality more easily.  Also updated the stretch
application code to use the get_normalized_band() function.

Refactored the update_display_image() code in rasterview.py to iterate
over the color bands, instead of duplicating code for each of
red/green/blue, and then again for grayscale.  That was gross.  Lost a
bit of readability because self._red_data etc. are now replaced by an
array self._display_data, but the code looked gross; I think this is a
bit cleaner.

The new code may be a bit slower than the old code.  This will be
addressed in future development efforts.