Linked sliders/limits, conditioner functions
Hooked up the "link sliders" and "link min/max values" check boxes to
properly gang up the controls of all channels.  This work needs to be
tested further; it is only active during RGB stretch building.

Hooked up the stretch and conditioner radio buttons to update the
stretch and conditioner on all the channels.  The histogram is updated
with the specified conditioner by applying it to the bin edges.