Implement more behavior of the new Stretch Builder UI
Refined a lot of interactions between the new UI widgets - min/max
fields, sliders, stretch low/high fields, general stretch-configuration
buttons, and "link xxxx" check-boxes.  The bulk of this functionality
is now working reasonably well.

Simplified/refined the object that represent stretch.  As the
abstraction becomes simpler, we may also replace this class hierarchy
with a simpler class hierarchy.

Generated new stretch/conditioner objects from the UI configuration,
so that the raster displays update when the user changes the stretch.
More work in this area is necessary; it is quite slow right now, and
we fire too many stretch-changed events when sliders are manipulated
by the user.

Refined the performance characteristics of the stretch builder UI.
Improved performance of histogram rendering, and constrained when
"stretch changed" events are fired to reduce the amount of display
updating.

Documented a lot of code too.