Initial documentation of contrast stretch
Wrote up a document describing how raster data is displayed, and how
contrast stretch is calculated and applied.  This is partly for user
documentation, and also partly for design review from the research team.

Also, the implementation currently reflects about 98% of the document;
a few implementation details deviate from it, and these deviations need
to be addressed in future fixes.