New spectrum plot features
The spectrum plot window is now able to show both single-pixel spectra
and area-average spectra.  Spectra are able to be collected, and are
shown in a tree-view, which will eventually include loaded spectral
libraries as well.  This tree-view also supports context menus, although
actions are not yet hooked up.

The spectrum plot window has its own configuration, for the default
behavior when the user clicks within the main window.  We may eventually
want to migrate this to a central configuration UI, but it also makes
sense to expose the UI-specific config in the UI.