Implement loading of ENVI spectral libraries
Created a SpectralLibrary abstraction that shares API similarities with
the RasterDataSet type.

Surprise surprise, GDAL doesn't support loading ENVI spectral library
files, even though they are also ENVI raster data files.  That was fun
to discover.

Migrated the proto1/envi.py data loading code into the ISWB codebase so
that we can use it to load ENVI spectral library files.  A few things
needed to be updated, but this was largely a very smooth update.

Implemented an ENVISpectralLibrary subclass of SpectralLibrary that can
load ENVI spectral library files.  It works when tested in the Python
command-line interface, so the next step will be to integrate it into
the spectral plot UI for some UI mocking.