Start integrating spectral libraries into app state
The demo for the spectral library / active / collected spectra had the
spectral state managed in the spectrum plot widget itself.  This is not
an appropriate architecture; the spectra should be maintained in the
overall application state.  This checkin starts to address this issue,
although it is clear that more efforts are necessary to keep the
app-state API from becoming unwieldy.