Move spectral library management to app state
Updated the spectral plot widget to register on the app-state class for
when spectral libraries are loaded and unloaded, to update the user
interface.  This allows us to load spectral libraries in multiple parts
of the UI, e.g. through the general "open file" UI.  This also allows us
to specify spectral library names on the command-line, to load them into
the workbench.

Factored out the application-state object from the application UI, since
that class is growing in complexity.