More spectrum context-menu functionality
Most of the context menu operations in the spectrum-list tree widget are
now functional.  This includes showing and hiding individual spectra,
showing and hiding all spectra in a group, unloading spectral libraries,
clearing the collected spectra, deleting individual spectra, and editing
the active or collected spectrum information.

The main feature that remains to be added is saving the collected
spectra to a file.