Refactor ROIs to remove rasterpane dependency
In preparation for supporting multiple raster-views in the main view
pane, dependencies directly on the rasterpane are being changed over to
raster-view dependencies.  This should allow for ROIs to be created in
one sub-view of the main view pane, and to be drawn in one sub-view of
the main view pane.