Restore basic zoom functionality to context/main panes
After introducing the ability to have multiple views in a raster pane,
a lot of the functionality in the main view and the context pane stopped
working.  This commit restores the zoom functionality, and gets the
basic zooming operations working again.

Drawing updates have not been verified yet, so displaying the viewport
regions etc. probably is still broken.  But, it is underway now.