Rework raster-pane toolbars; main-view management buttons
Raster-pane tool bar initialization now makes use of virtual functions
that subclasses can override, to provide other toolbar configurations
based on the raster pane's purpose.

Implemented a new "toolbar menu" widget that displays a list of
checkable options, and ensures that one option always remains checked.
This is basically the same code as the dataset chooser, so eventually
we should implement the dataset chooser in terms of this widget.  Not
critical to do right now though.

Added buttons for view management to the main raster pane, so that it
can be split and unsplit, and so that scrolling can be linked or
unlinked.  The behavior of these buttons is not yet implemented.