Hook up split view / link scroll buttons
The split/unsplit view button now updates the UI with the specified
structure of raster-views in the main raster pane.  Lots of stuff still
breaks when trying to display data, but at least it is updating the UI.

The link-scroll button is also hooked up, but scroll linking is not yet
done, since we can't even properly use all the raster views as yet.