Linked scrolling in main view, click actions in main/zoom views
This commit implements a simple scroll linking mechanism for the main
view when multiple raster-views are being displayed.  The scroll linking
can be toggled, of course.

Also began to re-wire up the clicking behaviors, so that clicking in
various raster-views would properly update the zoom pane and the
spectrum pane.  There are still some questions around this that need to
be nailed down.

The stretch builder and the band chooser still do not work yet.