Make selections, raster pane highlights data-set aware
Add data-set field to selections, so that they can be associated with
specific data when desirable to do so.

Updated rasterpane pixel-highlight code to use a SinglePixelSelection,
so that the pixel can also be associated with a data-set.

Also updated the rasterpane pixel-highlight drawing code to only draw
highlights on raster-views that are displaying the same dataset.  I'm
not sure if I like it, but now I can demo it.