Possible implementation of raster-view menubar (doesn't work)
This commit includes code for creating a menubar on each raster-view
when the main image window is showing multiple views.  It looks pretty
ugly, and I don't like it.  Leaving it in for the demo so we can discuss
it, but I'm hoping we can rip it out right away!!