Diagnostic and profiling functionality and docs
Added command-line options to support profiling the program, so that we
can identify performance issues.

Added faulthandler.enable() to the main entry point so that we can have
a better chance of diagnosing Qt/Pyside2 segfaults.

Added a diagnostics.md file with some documentation and code snippets
explaining what diagnostic and profiling options exist, and how to use
them.  This is very brief, and is barely even useful for developers,
let alone users.  Keep working on it.