Refactor/cleanup the raster-view image generation code
The purpose for refactoring and cleaning up this code is to make it
easier to access the image generation code directly from outside the
module.  Additionally, it should allow the code to be optimized for
performance (e.g. with Numba), etc.

This refactoring may later be undone, if it turns out that we really
need to cache the intermediate results to improve rendering performance.
However, I don't want to go down that path until we know we need it.

Secondarily, old code that is no longer being used is now eliminated.