Fix multi-view toolbars, stretch-building
I figured out that I left out one important line in getting the toolbars
to display in the raster-views when showing a multi-view display.  So
now I can forge ahead on making the multi-view toolbars.

I also updated the stretch-building machinery so that it will work with
multi-view displays.

I think we are ready for this branch to land in master.