New band chooser UI, fix stretch bugs
Implemented a new band chooser UI with the Qt Creator tool for the UI
layout.  Added the requested "choose default bands" button, and used
a stacked widget container to save a bit of space and to keep things
clearer for the user.  The band chooser also works with multi-views
now (although the interaction will undoubtedly need to be refined when
multiple views are showing).

While testing, a few stretch-builder bugs were encountered, having to do
with computing N% linear stretch on a data-set without a good data
distribution.  This is handled slightly better than before, but there
are many stretch-builder behaviors that still need ironing out.