Mac .app + .dmg build with PyInstaller
Makefile is updated to build a MacOS bundled application (.app) using
PyInstaller, and also a mountable disk-image file (.dmg) containing the
application.

It almost works - the application window comes up, but the icons are not
yet found by the app.  I don't want to lose this work, so checking it in!