Update Makefile with Windows app-freeze steps
Finally figured out how to run pyinstaller on Windows and generate a
frozen app that will pop up...  mostly.  Can't figure out how to build
a working *windowed* application, but at least it will run with a
console window.

The main issues were around ensuring that the proper Qt libraries are
included in the final package.  It is mostly straightforward to address
with --add-binary arguments.

The resources.py file must be regenerated per platform, so this needs
to be added to the Makefile.