Linked-views and stretch-builder button tweaks
A few small tweaks to the behavior of the stretch-builder button and the
linked-views button:

The stretch-builder button is only enabled if at least one data-set is
loaded.

The linked-views button is only enabled if at least two raster-views are
showing in the main raster pane.

The linked-views button will also pop up an error message if the
currently-loaded datasets aren't all the same dimensions.