Revised build process for generated Python files
Relocated the generated Python files to src/gui/generated so that
they're separate from the other source files.  Not comfortable
generating files into the same locations where hand-written sources
are.  See the directory's README.md for more details about what is
put in there.

Put together a couple of Makefiles to generate the resources into the
new directory.  Tied them all together at the top-level Makefile with a
"generated" target so that we can easily generate these files on any
platform we want to support.  This is necessary because the generated
files can vary across Qt versions, and one version may not be able to
load a file generated by another version.

Verified that you can do a "make clean build-mac" to generate a working
distributable from scratch.  Cool.