Fix ROI selection creation with multi-views
Region of Interest selections were only working in the first
raster-view when multiple views were enabled.  The bug turned out
to be an interesting issue in the way that task delegates have their
raster-views set; it was being set on any event from a raster-view.
This is problematic, because raster-views generate paint and
mouse-motion events all the time, and that was causing the wrong
raster-view to be set on the task delegate.

The fix is to only set the task delegate's raster-view when there is
intentional user interaction with a raster-view.  This means only mouse
press/release events, and key press/release events should cause a
delegate's raster-view to be set.

Once this fix was incorporated, the ROI selection creation code
started working as intended.  Hooray!