Rework rasterview for much more efficient image scaling
Previously, the rasterview would store a fully scaled version of the
image data being displayed.  This became prohibitive on extremely large
images, since 16x magnification is not uncommon.  The approach generated
a substantial performance penalty due to the extra work, and also a very
heavy memory burden from keeping the data around.

The rasterview's image display code has been reworked to keep only a
100% scale image in memory, and to leverage the Qt 5 optimized
draw-scaled-pixmap support.  The new version rescales much faster, and
the user interface is much more responsive.

The UI behavior is largely unchanged with this commit.  There are a few
niggling issues with views not being updated, but I think this commit is
of great enough value that it should be made available now, not after
those issues are addressed.