Region of Interest refactoring, usability improvements
The RegionOfInterest class is migrated into the raster package from the
gui package, since it is really a domain object, not GUI code.  This
also included splitting out some of the UI-specific ROI code and putting
it into the gui/ui_roi.py file.

Updated RegionOfInterest to support multiple selections, and also to
specify colors with a string, rather than a QColor.

Updated selection code to provide user guidance when selections are
being created.