Support for editing rectangle and polygon selections
The "edit geometry" context menu item now initiates editing of rectangle
and polygon selections.

The control-point size in the UI is now specified in a property, so that
all the editing code will draw similar results.

Polygon selection editing only allows the user to move around the
existing points of the polygon, not to add or delete points.