Improve selection-picking code accuracy, usability
There was a strange disparity between the accuracy of mouse-clicks and
context-menu clicks.  Ultimately this was due to the fact that
QMouseEvent objects can provide the click-coordinates as QPointF
(points with float coordinates), which made the "screen-coordinates" to
"dataset-coordinates" conversion more accurate.  QContextMenuEvent, on
the other hand, only provides click-coordinates as QPoint (points with
integer coordinates), which was causing loss of precision due to
rounding.  This was addressed in the raster-pane code.

Additional updates were made for managing selections:

The bounding box on multi-pixel selections was tweaked to ensure that
the right/bottom-most pixels in the selection would be included.

The picking of control-points when editing rectangle and polygon
selection was expanded so that it's easier to grab them; the user
doesn't need to be so precise, which makes things less frustrating.