Improve performance of ENVI spectral library loading
The ENVI spectral library code doesn't use memory-mapped IO since these
files are usually small.  However, the initialization code always
attempted to load the data file (not just the header file) before
verifying that the header's file-type indicated that it is actually a
spectral library.  This was awful when the data was a 30GB file.

Now the spectral library code will only open the ENVI header file and
verify the file type and structure before attempting to load any data.
Additionally, since we are assuming that spectral library files are
small, the code will log a warning if the spectral library's data size
is larger than a threshold (currently 10MiB).

Also went through and fleshed out ENVI code documentation, and added
type annotations.