Greatly improved Region of Interest support
Regions of Interest now support multiple selections, so that it's easy
to build up a multi-selection ROI.

When right-clicking on the raster display, the internal code identifies
the specific selections in the ROI that have been picked.  This allows
the context menu to provide items that can manipulate the various
sub-parts of the ROI.

The context menu now allows the the ROI's individual selections to be
edited or deleted, as well as the ROI information (name, color,
description) to be modified.

The context menu also exposees the ability to export the spectra of all
pixels in an ROI into a data file.

It is not yet possible to show the average spectrum of a Region of
Interest, although the code is completed to calculate what that average
spectrum would be.  Just need to figure out how to tie the components
together.