A few refinements to dataset information viewer
Got rid of the "General" group because it made the tree too deep, just
to see basic information about the dataset.

Added a count on the "Files" group so that it's easy to see how many
files comprise the data set.

Added an entry in the data-set info to indicate the "data ignore value,"
so that it's easy to diagnose potential issues with files missing this
value.