Unit testing framework and some tests
Started setting up the testing framework for this project.  Tests can be
run by typing "make test" in either the top level directory or the "src"
directory.

Test coverage is still almost nonexistent, but now that we have a
framework for it, it should be easier to add more tests in the future.