Factor internal state from spectrum plot into app state
Migrated the active/collected spectra from the spectrum plot window into
the main application state, so that it's easier for other components to
put spectra into the set of collected spectra.  This was somewhat
involved since I had to separate the non-GUI details of spectra from the
GUI/display details of spectra.

To facilitate this, I gave unique numeric IDs to individual spectra, and
to spectral libraries, so that the spectrum plot window can keep
separate "display information" for each spectrum being displayed.

The overall code is a bit more complicated now, since functionality is
spread over multiple classes.  One additional complexity is that all
spectral-plot display changes are now incremental, which makes the code
more challenging to understand, but it also results in a big performance
improvement for user interactions.  Unfortunately, the performance was
really bad before, so we need to live with this increased complexity.