Add (library_id, index) IDs to library spectra
Library spectra also need IDs to refer to them.  For now, it will be a
2-tuple of the form (library_id, index).

This conflicts with some of the type annotations which expect IDs to be
integers, but I will get back to patching all those things up.  This may
all change again when we support the Python interactive console.