Refactor spectrum info; export collected spectra
Migrated the SpectrumInfo class hierarchy to the raster directory, since
it now doesn't have hardly any GUI-specific code in it.  This required
updating a number of other Python files that reference these classes.

Finished the functionality for exporting spectra so that we can get
collected spectra and ROI pixel-spectra out of the program.  All of
this seems to work, and is hooked into the GUI now.