Add note to README.md about GDAL _gdal_array error
If the GDAL Python library is installed before NumPy is installed, there
will be subsequent errors saying that the _gdal_array module is not
present.  The issue is exacerbated due to the pip cache retaining the
non-NumPy-aware version of the GDAL Python library, so a special
argument needs to be passed to pip to make it all work properly.