Migrate Mac build to use a .spec file; add high-DPI display support
The packaged WISER Mac app wasn't displaying in a high-DPI manner on
high-DPI displays.  Needed to migrate the pyinstaller build process to
use a .spec file that includes some values for Info.plist that turns
this functionality on.  Also needed to turn on some options in main.py
to make this work.