Fix GDAL failure to load band on Windows/Miniconda3
On Windows and Miniconda3, the GDAL library doesn't seem to support the
use of virtual memory when retrieving a whole band of data from an
image.  This is concerning because some of these images are very large.
Nonetheless, the code now "falls back" to using non memory-mapped IO
when this error is encountered.