Support for exporting spectral plots as EPS/PNG/SVG/etc.
Integrated with matplotlib plot-export functionality to export spectral
plots to image files.  The dialog is simple at the moment, but gets the
job done.

The current default DPI is 100; should raise it to 300 or 600 for raster
formats like PNG.  Otherwise it looks very pixellated.

Hooks are also in for handling mouse-clicks on the spectral plot window.
That will require more work to complete though.