Config file + logging support for WISER app
Configuration file support for the WISER app is starting to be more
solid.  The config file can be specified as a command-line argument,
or a default config file can be loaded from default locations.

The main option for the config file is the ability to enable or disable
online bug reporting to BugSnag.  Several other useful options are
included in the config file, but this will continue to change over the
next few commits.

This commit also starts to incorporate the Python logging framework in
a more meaningful way.  Basic configuration is included.  There is
still much work to do on that front, though.