"Find display bands" uses first bands on dataset
If a dataset doesn't specify default bands, and if it doesn't have any
bands close to red/green/blue visible-light bands, it reports a tuple of
default bands to use.  Previously, it would choose first, middle and
last bands, but this didn't work for images of 3 or fewer bands.  Plus,
it just wasn't very good.

This commit changes the behavior to be much simpler:

*   If the data set has at least 3 bands, bands (0, 1, 2) are reported
    for the R/G/B color channels.

*   If the data set has fewer than 3 bands, band (0,) is reported as
    a tuple, for a grayscale channel.