Improve multi-view support for main rasterpane
This commit greatly improves the multi-view support in the main
rasterpane.  The raster-pane now includes a TiledRasterView subclass
of RasterView that adds a toolbar that includes the basic tools
necessary for managing a tiled array of raster-views.

The TiledRasterView toolbar includes a "dataset chooser" (a combobox),
a band chooser, and a stretch-builder for each window.  These tools are
only shown when more than one raster-view is shown in the pane.

When new data-sets are loaded into the application state, a raster-view
is chosen that is currently displaying no data.

Also, when the dimensions of the tiled pane are changed, the
raster-views that continue to be displayed, do not forget their internal
display state.