Update logic of linked/unlinked view updates
When in unlinked-views mode, clicks in the main/zoom pane will also
update the context pane (and the corresponding zoom/main pane) to show
the same dataset that was clicked on.  Pixel selections will also only
be indicated in a view showing the same dataset.

When in linked-views mode, clicks will not not change the dataset in
other views, and pixel selections are indicated in all main/zoom views.