Be smarter about showing datasets in main view in non-linked mode
In non-linked display mode, clicking in the zoom pane would always show
the same dataset in position (0,0) of the main viewing area - even when
it was already showing in a different pane.  Now the app will only show
the dataset in the main viewing area if it isn't already showing in some
view in the main area.