Click spectrum plot finds spectrum with y-value closest to click
The spectrum plot window now reacts to left mouse-clicks by finding the
spectrum with the y-value closest to the click.  A dotted vertical line
is drawn at the X-coordinate of the spectrum value.  If no spectrum is
found, the vertical line is hidden.

Next steps are to indicate the actual y-value of the intersection (e.g.
with a box), and to draw text somewhere to indicate the (X, Y) values.
Also, this functionality only works with spectra that have wavelengths;
need to do the same for spectra without wavelengths.