Display a point of interest on a spectrum plot
This is initial code for displaying a point of interest (i.e. from a
spectrum) on a spectrum-plot display.  It works pretty well.  The main
issue for now is that spectra can be updated - e.g. by changing their
average mode, or by changing their area, and the point doesn't update
to stay on the plot.  So, this needs to be fixed.