Updatable point-on-spectrum display for plot window
This version of the point-on-spectrum display code now supports updating
or removing the point-on-spectrum display when the spectrum itself
changes or is removed.  This is still a somewhat manual process, but it
is easier now that the class tracks the spectrum whose point is being
displayed.

The plot also shows the values being plotted now.