Initial support for configuration of major/minor ticks
The config dialog can now drive the inclusion/exclusion of major and
minor ticks on the X and Y axis.  This support is still very buggy, due
to the way that matplotlib specifies tick marks:  matplotlib expects a
list of numeric values specifying ticks.  This means that when the plot
changes the range being displayed in the X/Y axis, the tick positions
need to be recalculated.  This will need to be supported in a future
commit.  But, at least for now, the tick placement code seems to work.