Intermediate commit - spectrum plot config UI
This is an intermediate commit of the ongoing work for the spectrum plot
config UI.  The UI will break if you try to use most features on it.
However, the event-handling for checkboxes and their corresponding
line-edits, and validators on line-edits, are all hooked up, and this
seems a good point to commit.