Expose config options on spectrum plot; support legends
As the spectrum plot configuration UI becomes more complete, more
options are being exposed on the spectrum plot component, so the
dialog can access and control the component's configuration.

Support for enabling/disabling the plot legend is basically complete.
The legend also updates properly when plots change, are updated or
discarded.