Multiple updates to spectrum plot and config UI
Configuration of major and minor ticks has several usability
improvements.  The user interaction is more streamlined by the use of
group-boxes for the tick configuration, and when the user fills in an
interval then the UI will automatically enable the corresponding tick
option.  The UI also gives feedback about what the user can expect in
the tick UI.

Tick placement on the plot window works properly.  Automatic tick
placement still needs addressing.

Configuration options are added for tick-mark and selection font sizes.
These options still need to be fully implemented.

In the spectrum plot window, the click selection now supports click-drag
interactions, which will allow the selection point to be moved
incrementally left and right.  This code needs to be optimized at this
point.